Create Segmentation of Hamburg
In this project I would like to explore neighborhoods in Hamburg for availability of sports grounds and green areas. Such clustering and its visualisation can provide with an information about parts of the city with a lack of places to rest. This information can be used by management of the city to make Hamburg even more better place to live.  Moreover, this report is useful for someone who is looking for a place in Hamburg for recreational purposes.